Rename removeBeaber helper and update imports
