+++ src/helpers/removeBearer.ts
